No more args in main.rs
